requirements.txt
discord.py==2.3.2
fastapi==0.111.0
uvicorn[standard]==0.29.0
pyttsx3==2.90
numpy==1.26.4
requests==2.31.0
keyboard==0.13.5
PyNaCl==1.5.0
requests